refactor: split core completeness field editors and save strategies
The field registry was meant to be a readable table, but inline editor
JSX and inline save strategies had grown it to 497 lines, so scanning
"what fields exist" meant scrolling past dining type groups and JSON
parsing.

Split into:
- core-field.types.ts             CoreFieldConfig / CoreFieldEditorContext
- core-field-editors.tsx          all editors, incl. previously-inline
                                  type/phone/coordinates/idealFor/cuisines
- core-field-save-strategies.ts   all strategies, incl. previously-inline
- core-completeness-fields.tsx    the registry, now 171 lines

Pure extraction. Verified the registry's 24 keys and each entry's
draftInit/editor presence are identical to main.

Also: the two colocated test files in this directory do run under `bun
test` (9 tests), even though lm-client's package.json declares no test
runner and the files declare their own describe/expect stubs. They pass
identically on this branch and on main. Worth wiring up properly.

Verified: bun test 9/9 pass, same as main; tsc -p tsconfig.app.json
stays at 0 errors.
